chore: add all preprocessing checks

+++ examples/tests/preprocessing/loops/while-processed.c
@@ -1,4 +1,4 @@
-extern int snprintf(char *str, int size, const char *format, ...);
+extern int snprintf(char *str, unsigned long size, const char *format, ...);
 extern void klee_make_symbolic(void *addr, unsigned long nbytes, const char *name);
 extern void klee_assume(int condition);
 extern void klee_assert(int condition);
@@ -8,6 +8,9 @@ static char *__eclipse_int_to_string(int value, char *buffer, int buffer_size)
   snprintf(buffer, buffer_size, "%d", value);
   return buffer;
 }
+
+
+
 
 int __eclipse_original_main()
 {
@@ -26,4 +29,5 @@ int main(void)
 {
   return __eclipse_original_main();
 }
+
